create a build pipeline for the modular APIs
