+++ packages/app/next/package.json
@@ -1,0 +1,8 @@
+{
+    "name": "@firebase/app/next",
+    "main": "../dist/cjs/next.js",
+    "browser": "../dist/cjs/next.js",
+    "module": "../dist/esm5/next.js",
+    "esm2017": "../dist/esm2017/next.js",
+    "typings": "../ådist/packages/app/src/next/index.d.ts"
+}
